Basic implementation of ImportGames
Instead of providing metadata and possibly updating a game if it already exists, this gives us full control over how games are imported into the library. This update will allow all changes to games made in the server to be synced to the clients. It also simplifies the import, though it is more prone to exceptions due to the current use of yield.