Update Chunk Uploader and Import Dialog
The chunk uploader has been updated with a drag and drop interface as the old UI was always a bit temporary. The import dialog has been updated to support the new UI. Clearing now works and opening the import dialog for a second time does not have the previous file populated.